Agile Data Science/team_2_agile-main/requirements.txt
streamlit==1.15.0
urllib3==1.26.7
pandas == 1.3.3
numpy == 1.21.2
streamlit_lottie == 0.0.3
scikit-learn == 0.24.2
# h5py == 3.3.0
pyyaml
mlflow
